ios: Add autoreleasepool for generated images
When we generate a new image, we need to have a limited scope for the
generated image. Otherwise, it is put in the RPL thread, and never
released. This was leading to a black screen after a while running the
`Recurse` test, once we ran out of memory for new `UIImage`
instances.

Instrumentation using `CFGetRetainCount` showed that 2 references
typically remained when we were looking at the old image being
replaced in the `ScreenView`. I do not understand where they live,
but I know that adding `@autoreleasepool` allows the compiler to
figure it out.

Moral of the story: Objective-C ARC is not C++ smart pointers.
There is a garbage collection component to it.

See more explanations here:
https://stackoverflow.com/questions/25101154/how-to-deal-with-arc-in-a-background-thread

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>