ios: Rework file selector so that it confirms overwriting
There is some really "smart" semantics in the iOS file selector, which
makes it duplicate existing files by default if you pass it a writable
file as input (i.e. the last file you loaded).

On the other hand, if you pass a bundle file, then it asks if you want
to overwrite, which is the desired behaviour.

So we always pass the `state/Demo.48S` file, which is expected to
exist in all cases. A bit weird, but does the job.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>