library: Add cache for library objects to accelerate evaluation
In order to ensure that library items are evaluated rapidly, keep a
cache of loaded / compiled library objects above the directories.

The performance of this optimization can be evaluated by comparing the
Collatz Conjecture benchmark in its library implementation before or
after, and compare to the global variable version.

DM42 on USB:
- Globals 15.6s
- Library before: 75.2s
- Library after: 15.5s

Simulator:
- Globals: 1.9s
- Library before: 2.9s
- Libarary after: 1.6s

Fixes: #1217

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>